Crate bridge_adapters

Modules

lisp_adapters

My notes: #. To convert a rust value to an slosh type impl SlFrom<T> for Value, #. To convert a slosh &Value to an owned type implement impl SlFromRef<'a, &Value> for RustType, this allows rust native functions annotated with the bridge macro to receive normal rust types. #. To convert a slosh &Value to a mutable reference type implement impl SlAsMut<&Value> for MutRefType. #. To convert some rust type back to a value that the rust native function annotated by the bridge macro returns implement impl SlFrom<&Value> for RustType. TODO PC ISSUE #7 - returning values via annotated or lifetime? To avoid allocations when converting a slosh &Value back to a rust type that was mutated don’t return anything. If it is necessary for the API to return some value. …either use an annotation on an input argument fn myfun(#[likethis] returnme: &mut String, someotherval: String) -> VMResult<()> or a lifetime… might be easier to do the annotation.

Functions

add_builtin