Module bridge_adapters::lisp_adapters

My notes: #. To convert a rust value to an slosh type impl SlFrom<T> for Value, #. To convert a slosh &Value to an owned type implement impl SlFromRef<'a, &Value> for RustType, this allows rust native functions annotated with the bridge macro to receive normal rust types. #. To convert a slosh &Value to a mutable reference type implement impl SlAsMut<&Value> for MutRefType. #. To convert some rust type back to a value that the rust native function annotated by the bridge macro returns implement impl SlFrom<&Value> for RustType. TODO PC ISSUE #7 - returning values via annotated or lifetime? To avoid allocations when converting a slosh &Value back to a rust type that was mutated don’t return anything. If it is necessary for the API to return some value. …either use an annotation on an input argument fn myfun(#[likethis] returnme: &mut String, someotherval: String) -> VMResult<()> or a lifetime… might be easier to do the annotation.

rosetta stone for bridge macros

Rust Type	Slosh Type & Traits

S -> R Convert Slosh -> Rust
- Occurs when coercing slush arguments to the parameter types in the signature of the annotated Rust function.
R -> S Convert Rust -> Slosh
- Occurs when coercing some returned Rust type to a Slosh type.

String	Value::String

S -> R

- SlIntoRef String for &Value

R -> S

- SlFrom &Value for String

&String	Value::String

S -> R

- SlIntoRef &String for &Value

R -> S

- take String

* uses Clone unless TODO PC ISSUE #7 the extant value problem

&mut String	Value::String

S -> R

- SlAsMut String for &Value

R -> S

- take &mut String

* uses Clone unless TODO PC ISSUE #7 the extant value problem

&str	Value::String / Value::StringConst

S -> R

- SlAsRef &str for &Value

- SlIntoRef &str for &Value

R -> S

- SlFrom for Value

* uses Clone unless TODO PC ISSUE #7 - the extant value problem

- TODO PC ISSUE #7 adjacent is it even possible to call vm.alloc_string_ro on something that was newly created in the current fcn and returned as a RO value OR should that be made as a custom type so the user can declare their intent.

char	Value::CodePoint

S -> R

- SlIntoRef char for &Value

R -> S

- SlFrom &Value for char

SloshChar	Value::CharClusterLong / Value::CharCluster / Value::CodePoint

S -> R

- SlIntoRef SloshChar for &Value

R -> S

- SlFromRef &Value for SloshChar

LooseString	Value::String / Value::CodePoint / Value::CharCluster / Value::CharClusterLong / Value::Symbol / Value::Keyword / Value::StringConst

S -> R

- SlIntoRef LooseString for &Value

R -> S

* Note: Always does an allocation and returns a Value::String type.

- SlFromRef &Value for LooseString

bool	Value::True / Value::False / Value::Nil

S -> R

- SlIntoRef bool for &Value

R -> S

- SlFrom &Value for primitives

Value::Byte(u8),

Value::Int([u8; 7]), // Store a 7 byte int (i56…).

Value::Float(F56),

Value::List(Handle, u16), Value::Vector(Handle), Value::Map(Handle),

Value::Special(Interned), // Intended for symbols that are compiled. Value::Builtin(u32), Value::Undefined, Value::Nil,

Modules

numbers

primitives

text

Traits

SlAsMut

SlAsRef

Converts a BridgedType to some rust type

SlFrom

Use mutable SloshVm to take a rust value and convert it to a BridgedType.

SlFromRef

SlFromRefMut

SlInto

Inverse of SlFrom

SlIntoRef

SlIntoRefMut